Add files via upload
This zipped folder has two locations, Pershing state park site and Saxton site which are located in Linn county in Missouri. the inputs are already installed inside the geodatabase file with the script inside it. the zipped folder should be unzipped and the path for the location of this file should be entered inside the Python code.